Code Review update. Update Examples to say halting rather than spinning. Add comments about why break time is set. Move driver to it's own underlying file. Make getDeviceID change op mode if needed. Add checking to setDeviceAddress. Make runDefaultSetup change op mode if needed. Move bulk of runDefaultSetup into the else statement so we don't do unnecessary things. prepareMeasurement now uses references and consts. Add comment to reset about OSR register availability. Add human readable versions of getters for gain, cclk, convTime. Rename original versions of those to Raw. Remove unnecessary if's in readRawUV. setDigitalDividerRange now defaults enabling the divider to true. Modify Arduino begin to take in a wire port. Remove unnecessary reset call in begin due to logic changes above.
