+++ examples/Example01_Basic_OneShot/Example01_Basic_OneShot.ino
@@ -38,7 +38,7 @@ void setup() {
   // Initialize sensor and run default setup.
   if(myUVSensor.begin() == false) {
     Serial.println("Sensor failed to begin. Please check your wiring!");
-    Serial.println("Spinning...");
+    Serial.println("Halting...");
     while(1);
   }
 
@@ -47,7 +47,7 @@ void setup() {
   // Set measurement mode and change device operating mode to measure.
   if(myUVSensor.prepareMeasurement(MEAS_MODE_CMD) == false) {
     Serial.println("Sensor did not get set properly.");
-    Serial.println("Spinning...");
+    Serial.println("Halting...");
     while(1);
   }
 
@@ -61,7 +61,7 @@ void loop() {
     Serial.println("Error starting reading!");
   
   // Wait for a bit longer than the conversion time.
-  delay(2+(1 << myUVSensor.getConversionTime()));
+  delay(2+myUVSensor.getConversionTimeMillis());
 
   // Read UV values.
   if(kSTkErrOk != myUVSensor.readAllUV())
