+++ examples/Example03_SYNS_Mode/Example03_SYNS_Mode.ino
@@ -53,7 +53,7 @@ void setup() {
   // Initialize sensor and run default setup.
   if(myUVSensor.begin() == false) {
     Serial.println("Sensor failed to begin. Please check your wiring!");
-    Serial.println("Spinning...");
+    Serial.println("Halting...");
     while(1);
   }
 
@@ -62,7 +62,7 @@ void setup() {
   // Set measurement mode and change device operating mode to measure.
   if(myUVSensor.prepareMeasurement(MEAS_MODE_SYNS) == false) {
     Serial.println("Sensor did not get set properly.");
-    Serial.println("Spinning...");
+    Serial.println("Halting...");
     while(1);
   }
